+++ docs/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.6] - 2026-08-12
+
+### Added
+
+- Accept a double-quoted attribute values, as in `path="/path/with spaces/FILE.md"`
+
+
 ## [0.0.5] - 2026-08-11
 
 ### Changed